+++ GameEngine/CoreEngine/CoreEngine/src/Coordinates.h
@@ -13,8 +13,8 @@ namespace Engine
 		int Z = 0;
 
 		constexpr Coordinates(int x = 0, int y = 0, int z = 0) : X(x), Y(y), Z(z) {}
-		Coordinates(const Vector3& vector) : X(int(std::floorf(vector.X))), Y(int(std::floorf(vector.Y))), Z(int(std::floorf(vector.Z))) {}
+		Coordinates(const Vector3& vector) : X(int(std::floor(vector.X))), Y(int(std::floor(vector.Y))), Z(int(std::floor(vector.Z))) {}
 
-		operator Vector3() const { return Vector3(float(X), float(Y), float(Z)); }
+		operator Vector3() const { return Vector3(Float(X), Float(Y), Float(Z)); }
 	};
 }
